requirements.txt
langchain
langchain_chroma
langchain-openai
chromadb
pypdf
tiktoken
python-dotenv
cryptography
streamlit